<refactor> Parser 관련 패키지 구조 수정
