+++ src/main/java/webserver/utils/parser/body/RequestBodyParser.java
@@ -1,4 +1,4 @@
-package webserver.utils.parser;
+package webserver.utils.parser.body;
 
 import java.util.Map;
