+++ src/main/java/webserver/utils/parser/body/RequestBodyParserManager.java
@@ -1,4 +1,4 @@
-package webserver.utils.parser;
+package webserver.utils.parser.body;
 
 import webserver.exceptions.BadRequestException;
